requirements.txt
aioredis==2.0.1
alembic==1.14.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.2.1
billiard==4.2.1
black==25.1.0
celery==5.4.0
cfgv==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distlib==0.3.9
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.8
filelock==3.17.0
flake8==7.1.2
greenlet==3.1.1
h11==0.14.0
identify==2.6.8
idna==3.10
isort==6.0.1
kombu==5.4.2
Mako==1.3.9
MarkupSafe==3.0.2
mccabe==0.7.0
mypy==1.15.0
mypy-extensions==1.0.0
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
pre_commit==4.1.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycodestyle==2.12.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyflakes==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.2.1
rsa==4.9
setuptools==65.5.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
vine==5.1.0
virtualenv==20.29.2
wheel==0.38.4
